requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
chromadb==0.4.17
requests==2.31.0
pytest==7.4.3
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.4
pandas==2.1.3
python-dotenv==1.0.0
huggingface-hub==0.13.3
openai==1.8.3